Struct ldconvert::lookup::LookupTable

pub struct LookupTable {
    name: String,
    values: Vec<Vec<Vec<Oklab>>>,
    resolution: usize,
    palette: HashMap<u32, Oklab>,
    kernel: Vec<KernelEntry>,
}

Generates a LUT based on a palette, and does color quantization for convertable images.

Fields
name: String

Used for the naming of the save file.

values: Vec<Vec<Vec<Oklab>>>resolution: usize

LUT resolution.

The number of steps that each color channel is broken into.

For the png format, 255 covers all the possible colors.

palette: HashMap<u32, Oklab>kernel: Vec<KernelEntry>

Diffusion kernel. Stores the weights used to distribute the error value over nearby colors when quantizing the LUT.

Implementations

impl LookupTable

pub fn new(palette_path: PathBuf, resolution: usize) -> LookupTable

Initializes fields and fills the kernel vector.

Kernel weights are distributed according to the Minimized Average Error algorithm, extended to the third dimension. This divides the error into 108 parts.

pub fn populate(self) -> Self

Fills the LUT with clean RGB values.

pub fn discretize(self) -> Self

Quantizes the LUT with error diffusion.

fn discretize_color(palette: &HashMap<u32, Oklab>, color: &Oklab) -> Oklab

Quantizes a color to the nearest color to it in the palette.

Uses Oklab euclidean distance for color difference calculation.

fn scale(color: Oklab, scale: f32) -> Oklab

Scales the color components.

Used for error diffusion, result otherwise nonsensical.

pub fn lookup(&self, color: Oklaba) -> Oklaba

Applies the LUT to a convertible image color.

The color is rounded to the nearest valid LUT step, in case of a low resolution LUT.

pub fn save_slices(&self, save_path: PathBuf)

Saves slices of the lookup table as images using the provided save path.

Slices are used for testing the validity and coverage of the conversion.

pub fn save_lut(&self, save_path: PathBuf)

Saves the LUT as a binary file using the provided save path.